+++ spring-security-jwt/pom.xml
@@ -9,7 +9,7 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.demo</groupId>
-	<artifactId>spring-security-jwt</artifactId>
+	<artifactId>fisPlanOnboarding</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>spring-security-jwt</name>
 	<description>Demo project for User Login</description>
