+++ spring-security-jwt/src/main/java/com/fis/deloitte/planOnboarding/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.demo.spring_security_jwt.config;
+package com.fis.deloitte.planOnboarding.config;
 
 import org.springframework.context.annotation.Configuration;
